requirements.txt
anyio==3.7.1
APScheduler==3.10.4
certifi==2023.7.22
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.4
python-telegram-bot==20.4
pytz==2023.3
six==1.16.0
sniffio==1.3.0
tzlocal==5.0.1
